functions/requirements.txt
anthropic
requests
beautifulsoup4
pyyaml
python-dotenv
functions-framework==3.*
firebase-admin
google-cloud-firestore
